Added the rest of the possibilities to the at_bat_class
